pyproject.toml
[project]
name = "excel-grademaster-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.56.0",
    "openpyxl>=3.1.2",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
]
